requirements/base.txt
Django==1.7
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
psycopg2==2.5.4
djangorestframework==2.4.3
djangorestframework-gis==0.7
simplejson==3.6.3
